requirements.txt
streamlit>=1.31.0
numpy>=1.24.0
pandas>=2.0.0
requests>=2.31.0
PyPDF2>=3.0.0
PyMuPDF>=1.23.0
python-docx>=0.8.11
pytesseract>=0.3.10
Pillow>=10.0.0
opencv-python>=4.8.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
scikit-learn>=1.3.0
langchain>=0.1.0
langchain-text-splitters>=0.0.1


